Fix manifest type factory instantiation and discovery
I swear this worked previously but today I had issues with the
linter telling me my objects couldn't be instantiated and my
types were bad. It is correct, they definitely were wrong. So I
changed the map to store the types as any, created a new
manifest factory, and updated the discover method to
use the new discovery method.