Add commander and a command line interface
I had to, for lack of knowledge of a faster option, create 2
new interfaces off a base interface. User options come from
commander / argv where application options are what are required
to properly execute.

The getOptionsWithDefaults takes in a partial set of IOptions
(to handle both commander or direct use of IOptions) partial and
returns all fields required of IApplication options. A manifest
object is created from the manifest file if one is provided.

The program continues from there normally.